Very basic crop implementation
Adds a very basic crop implementation using canvas.
This crop is based on the selection rectangle only, so this removes the
display of the various controls so users will just move the selection
rectangle to define the crop.

The cropped image is then output to $scope.resultImage (i.e.
result-image='...' in the element).

Test Plan:
Use the test page and drag the rectangle around, ensuring the sample
image looks correct for the selection.